+++ static/app/utils/metrics/useMetricsSamples.tsx
@@ -1,5 +1,7 @@
 import {normalizeDateTimeParams} from 'sentry/components/organizations/pageFilters/parse';
 import type {MRI} from 'sentry/types';
+import {defined} from 'sentry/utils';
+import {parseMRI} from 'sentry/utils/metrics/mri';
 import {useApiQuery} from 'sentry/utils/queryClient';
 import useOrganization from 'sentry/utils/useOrganization';
 import usePageFilters from 'sentry/utils/usePageFilters';
@@ -13,7 +15,7 @@ interface UseMetricSamplesOptions<F extends string> {
   query?: string;
 }
 
-interface MetricSamplesResults<F extends string> {
+export interface MetricsSamplesResults<F extends string> {
   data: {
     [K in F]: string[] | string | number | null;
   }[];
@@ -46,10 +48,43 @@ export function useMetricsSamples<F extends string>({
     },
   };
 
-  return useApiQuery<MetricSamplesResults<F>>([path, endpointOptions], {
+  return useApiQuery<MetricsSamplesResults<F>>([path, endpointOptions], {
     staleTime: 0,
     refetchOnWindowFocus: false,
     retry: false,
     enabled,
   });
+}
+
+export function isSupportedMRI(mri: MRI): boolean {
+  // extracted transaction metrics
+  if (mri === 'd:transactions/duration@millisecond') {
+    return true;
+  }
+
+  // extracted span metrics
+  if (
+    mri === 'd:spans/exclusive_time@millisecond' ||
+    mri === 'd:spans/duration@millisecond'
+  ) {
+    return true;
+  }
+
+  const parsedMRI = parseMRI(mri);
+  if (defined(parsedMRI)) {
+    // extracted measurement metrics
+    if (
+      parsedMRI.useCase === 'transactions' &&
+      parsedMRI.name.startsWith('measurements.')
+    ) {
+      return true;
+    }
+
+    // user defined custom metrics
+    if (parsedMRI.useCase === 'custom') {
+      return true;
+    }
+  }
+
+  return false;
 }
